Revert FormLayerComponent mapping to original file
