+++ pages/video-editor/components/layers/FormLayer.js
@@ -17,7 +17,77 @@ import { updateLayerField } from '../../redux/slice/videoSlice';
 import ColorPickerButton from '../shared/color-picker/ColorPickerButton.jsx';
 import LayersHeader from './LayersHeader.js';
 import AjaxWarning from '../forms/AjaxWarning.js';
-import { FormLayerComponentType, getFormIdFromLayer } from '../../utils/formUtils';
+import { getFormIdFromLayer } from '../../utils/formUtils';
+import GravityForm from '../forms/GravityForm';
+import WPForm from '../forms/WPForm';
+import EverestForm from '../forms/EverestForm';
+import CF7 from '../forms/CF7';
+import JetpackForm from '../forms/JetpackForm';
+import SureForm from '../forms/Sureform.js';
+import FluentForm from '../forms/FluentForm.js';
+import ForminatorForm from '../forms/forminatorForms.js';
+
+/**
+ * FormLayer Components Object mapping.
+ */
+export const FormLayerComponentType = {
+	gravity: {
+		isActive: Boolean( window?.videoData?.gfActive ) ?? false,
+		component: GravityForm,
+		idField: 'gf_id',
+		settingsUrl: 'admin.php?subview=confirmation&page=gf_edit_forms&id={formId}&view=settings',
+	},
+	cf7: {
+		isActive: Boolean( window?.videoData?.cf7Active ) ?? false,
+		component: CF7,
+		idField: 'cf7_id',
+		settingsUrl: 'admin.php?page=wpcf7&post={formId}&action=edit',
+	},
+	jetpack: {
+		isActive: Boolean( window?.videoData?.jetpackActive ) ?? false,
+		component: JetpackForm,
+		idField: 'jp_id',
+		settingsUrl: 'admin.php?page=jetpack-forms-admin#/responses',
+		// Special handling for Jetpack forms (extract post ID from form ID)
+		getFormId: ( formId ) => {
+			if ( ! formId ) {
+				return null;
+			}
+			const parts = formId.split( '-' );
+			return parts[ 0 ] ? parseInt( parts[ 0 ] ) : null;
+		},
+	},
+	wpforms: {
+		isActive: Boolean( window?.videoData?.wpformsActive ) ?? false,
+		component: WPForm,
+		idField: 'wpform_id',
+		settingsUrl: 'admin.php?page=wpforms-builder&view=settings&form_id={formId}&section=general',
+	},
+	sureforms: {
+		isActive: Boolean( window?.videoData?.sureformsActive ) ?? false,
+		component: SureForm,
+		idField: 'sureform_id',
+		settingsUrl: 'post.php?post={formId}&action=edit',
+	},
+	forminator: {
+		isActive: Boolean( window?.videoData?.forminatorActive ) ?? false,
+		component: ForminatorForm,
+		idField: 'forminator_id',
+		settingsUrl: 'admin.php?page=forminator-cform-wizard&id={formId}',
+	},
+	fluentforms: {
+		isActive: Boolean( window?.videoData?.fluentformsActive ) ?? false,
+		component: FluentForm,
+		idField: 'fluent_form_id',
+		settingsUrl: 'admin.php?page=fluent_forms&form_id={formId}&route=settings&sub_route=form_settings',
+	},
+	everestforms: {
+		isActive: Boolean( window?.videoData?.everestFormsActive ) ?? false,
+		component: EverestForm,
+		idField: 'everest_form_id',
+		settingsUrl: 'admin.php?page=evf-builder&view=fields&form_id={formId}&tab=settings',
+	},
+};
 
 /**
  * Component to render and manage a form layer within the video editor.
